+++ tests/test_events.py
@@ -3,19 +3,21 @@ from __future__ import division
 from __future__ import print_function
 from __future__ import unicode_literals
 
+from datetime import datetime
 from copy import deepcopy
 
 import pytest
 
-from rasa_core.events import UserUttered, TopicSet, SlotSet, Restarted, \
-    ActionExecuted, AllSlotsReset, \
-    ReminderScheduled, ConversationResumed, ConversationPaused, StoryExported, \
-    ActionReverted, BotUttered
+from rasa_core.events import (
+    Event, UserUttered, TopicSet, SlotSet, Restarted,
+    ActionExecuted, AllSlotsReset,
+    ReminderScheduled, ConversationResumed, ConversationPaused,
+    StoryExported, ActionReverted, BotUttered)
 
 
 @pytest.mark.parametrize("one_event,another_event", [
-    (UserUttered("/greet", {"intent": "greet", "confidence": 1.0}, []),
-     UserUttered("/goodbye", {"intent": "goodbye", "confidence": 1.0}, [])),
+    (UserUttered("/greet", {"name": "greet", "confidence": 1.0}, []),
+     UserUttered("/goodbye", {"name": "goodbye", "confidence": 1.0}, [])),
 
     (TopicSet("my_topic"),
      TopicSet("my_other_topic")),
@@ -68,3 +70,36 @@ def test_event_has_proper_implementation(one_event, another_event):
     # str test
     assert "object at 0x" not in str(one_event), \
         "Event has a proper str method"
+
+
+@pytest.mark.parametrize("one_event", [
+    UserUttered("/greet", {"name": "greet", "confidence": 1.0}, []),
+
+    TopicSet("my_topic"),
+
+    SlotSet("name", "rasa"),
+
+    Restarted(),
+
+    AllSlotsReset(),
+
+    ConversationPaused(),
+
+    ConversationResumed(),
+
+    StoryExported(),
+
+    ActionReverted(),
+
+    ActionExecuted("my_action"),
+
+    BotUttered("my_text", "my_data"),
+
+    ReminderScheduled("my_action", datetime.now())
+])
+def test_dict_serialisation(one_event, default_domain):
+    evt_dict = one_event.as_dict()
+    recovered_event = Event.from_parameters(one_event.type_name,
+                                            evt_dict,
+                                            default_domain)
+    assert hash(one_event) == hash(recovered_event)